reorganized some functions between files!
